Add multiple stages as a best practice [DE-7]
Stages are a level of isolations: you cannot access information form other stages, but stacks can use informations from other stacks. Defining all stages in app.py is also best practice for how to parametrize the stack.

Deployments now need to include the stage/stack name to work. The makefile is per default using the dev stage to do this.

This also splits a bit synth from deploy: running deploy without synth beforehand still works, but if there is already an out dir, that is used. This was also recommended as a best practice to not deploy stuff which hasn't been tested (although I do not see how that dir is tested at all)